Renamed BedService to Cas1BedService and moved to cas1 package.
